refactor: refactor computeType handling for platform compatibility
- Import the status module in processAudio.ts
- Update computeType values based on the platform in status.ts
- Change computeType for win32 from int8 to float16
- Change computeType for darwin from float16 to int8
- Change computeType for linux from float32 to float16

+++ src/routes/status.ts
@@ -17,15 +17,15 @@ router.get('/status', (_req: Request, res: Response) => {
   switch(platform) {
     case 'win32':
       osName = 'windows';
-      computeType = 'int8';
+      computeType = 'float16';
       break;
     case 'darwin':
       osName = 'mac';
-      computeType = 'float16';
+      computeType = 'int8';
       break;
     case 'linux':
       osName = 'linux';
-      computeType = 'float32';
+      computeType = 'float16';
       break;
     default:
       osName = platform;
